requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
openai==1.33.0
azure-identity==1.16.1
python-dotenv==1.0.1
environs==11.0.0
aiohttp==3.9.5
loguru==0.7.2
azure-search-documents==11.6.0b4
azure-storage-blob==12.21.0
